requirements.txt
utm==0.7.0
shapely==1.8.4
typer[all]==0.6.1
python-decouple==3.6
jupyter==1.0.0
pandas==1.5.0
plotly==5.10.0